update codes, upgrade BLINKER APCONFIG codes
